fix: 로그인 안했을 때 answer 페이지에서 좋아요 눌러지는 현상을 해결

+++ src/pages/Answer.tsx
@@ -68,8 +68,6 @@ const Answer = () => {
   const [question, setQuestion] = useState<string>(initialQuestion);
   const [content, setContent] = useState<string>(initialContent);
 
-  console.log(submittedAnswers);
-
   const navigate = useNavigate();
   useEffect(() => {
     const fetchAnswers = async () => {
@@ -107,6 +105,11 @@ const Answer = () => {
 
   const handleLike = async (id: string) => {
     try {
+      const userData = localStorage.getItem("userData");
+      if (!userData) {
+        alert("로그인이 필요합니다.");
+        return;
+      }
       const likeDocRef = doc(db, "likes", `${currentUser.uid}_${id}`);
       if (likedPosts.has(id)) {
         await deleteDoc(likeDocRef);
